fix(cli): resolve MSVC C4244 narrowing in catalog sync progress logging
FormatDuration() takes `long`, but the callback at command_executor.cpp:2744
passed `elapsed_s` (an auto-deduced std::chrono::seconds::rep, i.e. int64_t)
straight through. On Windows (LLP64: `long` is 32-bit) MSVC's /W4 /WX treats
the implicit narrowing as a hard error; on Linux/macOS (LP64: `long` is
64-bit) it's a no-op conversion, so this only broke the Windows build. Cast
explicitly with static_cast<long>, matching the pattern already used for
`eta_s` two lines above. Verified with a local g++ -Wconversion sweep of the
whole file — no other instances.

+++ src/cli/command_executor.cpp
@@ -2741,8 +2741,8 @@ int HandleCatalogSync(const CommandArgs& args) {
             ? static_cast<long>(elapsed_s) * (p.total - p.index) / p.index
             : 0;
         std::cerr << "[" << p.index << "/" << p.total << "] " << p.item_kind << " " << p.item_name
-                  << " (elapsed " << FormatDuration(elapsed_s) << ", ETA " << FormatDuration(eta_s)
-                  << ")\n";
+                  << " (elapsed " << FormatDuration(static_cast<long>(elapsed_s)) << ", ETA "
+                  << FormatDuration(eta_s) << ")\n";
     };
 
     auto result = CatalogSync(*session, *store, *options, scope_label, pipeline);
